requirements.txt
PyYAML==3.12
gitpython==2.0.8
termcolor==1.1.0